+++ src/core/drivers/index.ts
@@ -2,4 +2,5 @@ export { DatabaseDriver, BaseDriver } from './DatabaseDriver';
 export { MySQLDriver } from './MySQLDriver';
 export { PostgreSQLDriver } from './PostgreSQLDriver';
 export { SQLiteDriver } from './SQLiteDriver';
+export { ClickHouseDriver, normalizeClickHouseType } from './ClickHouseDriver';
 export { DriverFactory } from './DriverFactory';